@@ -1,50 +1,50 @@
 {
-    "name": "rahulbhooteshwar.github.io",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build --prod --output-hashing all",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^5.2.0",
-        "@angular/common": "^5.2.0",
-        "@angular/compiler": "^5.2.0",
-        "@angular/core": "^5.2.0",
-        "@angular/forms": "^5.2.0",
-        "@angular/http": "^5.2.0",
-        "@angular/platform-browser": "^5.2.0",
-        "@angular/platform-browser-dynamic": "^5.2.0",
-        "@angular/router": "^5.2.0",
-        "core-js": "^2.4.1",
-        "lodash": "^4.17.5",
-        "ngx-busy-loader": "^2.1.1",
-        "rxjs": "^5.5.6",
-        "zone.js": "^0.8.19"
-    },
-    "devDependencies": {
-        "@angular/cli": "~1.7.1",
-        "@angular/compiler-cli": "^5.2.0",
-        "@angular/language-service": "^5.2.0",
-        "@types/jasmine": "~2.8.3",
-        "@types/jasminewd2": "~2.0.2",
-        "@types/node": "~6.0.60",
-        "codelyzer": "^4.0.1",
-        "jasmine-core": "~2.8.0",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~2.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "^1.2.1",
-        "karma-jasmine": "~1.1.0",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "~5.1.2",
-        "ts-node": "~4.1.0",
-        "tslint": "~5.9.1",
-        "typescript": "~2.5.3"
-    }
+  "name": "my-profile-upgrade",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --open",
+    "build": "ng build --prod --output-hashing all",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/router": "~7.2.0",
+    "core-js": "^2.5.4",
+    "lodash": "^4.17.5",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
+    "typed.js": "^2.0.10",
+    "zone.js": "~0.8.26"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "~7.3.9",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
+  }
 }
